Peakprep a

+++ server/src/controllers/adminControls.ts
@@ -15,7 +15,7 @@ export const addHr = async (req: Request, res: Response): Promise<any> => {
     // username <-> email, password, role: hr, name, company, 
     const { email, password, name, company } = req.body;
     if (!email || !password || !name || !company) {
-        return res.status(400).send({ ...response, error: 'Missing required fields.' });
+        return res.status(400).json({ ...response, error: 'Missing required fields.' });
     }
     let newUser: IUser | null = null;
     let newHr: IHR | null = null;
@@ -24,11 +24,11 @@ export const addHr = async (req: Request, res: Response): Promise<any> => {
         const existingUser = await User.findOne({ username: email });
 
         if (existingUser || existingHr) {
-            return res.status(400).send({ ...response, error: 'Email already exists for another user.' });
+            return res.status(400).json({ ...response, error: 'Email already exists for another user.' });
         }
         const companyRecord = await Company.findOne({ name: company });
         if (!companyRecord) {
-            return res.status(404).send({ ...response, error: 'Company not found.' });
+            return res.status(404).json({ ...response, error: 'Company not found.' });
         }
         const userId = uuidv4();
         const hashedPassword = await bcrypt.hash(password, 10);
@@ -41,7 +41,7 @@ export const addHr = async (req: Request, res: Response): Promise<any> => {
         await companyRecord.save();
 
         const { userId: hrUserId, ...hrDetails } = newHr.toObject();
-        res.send({ ...response,success: true, message: 'HR added successfully.', data: { hr: hrDetails } });
+        res.status(201).json({ ...response, success: true, message: 'HR added successfully.', data: { hr: hrDetails } });
     } catch (error) {
         if (newUser) {
             await newUser.deleteOne();
@@ -49,19 +49,19 @@ export const addHr = async (req: Request, res: Response): Promise<any> => {
         if (newHr) {
             await newHr.deleteOne();
         }
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
 export const deleteHr = async (req: Request, res: Response): Promise<any> => {
     const { email } = req.body;
      if (!email) {
-        return res.status(400).send({ ...response, error: 'Email required' });
+        return res.status(400).json({ ...response, error: 'Email required' });
     }
     try {
       const hr = await HR.findOne({ email });
         if (!hr) {
-            return res.status(404).send({ ...response, error: 'HR not found.' });
+            return res.status(404).json({ ...response, error: 'HR not found.' });
         }
         const company = await Company.findById(hr.companyId);
         if (company) {
@@ -70,16 +70,16 @@ export const deleteHr = async (req: Request, res: Response): Promise<any> => {
         }
         await User.findOneAndDelete({ username: email });
         await HR.findOneAndDelete({ email });
-        res.send({ ...response, success: true, message: 'HR deleted successfully.', hr });
+        res.status(200).json({ ...response, success: true, message: 'HR deleted successfully.', data: hr });
     } catch (error) {
-        res.status(500).send({ ...response, error: 'Internal server error.' });
+        res.status(500).json({ ...response, error: 'Internal server error.' });
     }
 }
 
 export const addStudent = async (req: Request, res: Response): Promise<any> => {
     const { username, password, name, branch, section, admissionYear } = req.body;
     if ( !password || !name || !branch || !section || !admissionYear) {
-        return res.status(400).send({ ...response, error: 'Missing required fields.' });
+        return res.status(400).json({ ...response, error: 'Missing required fields.' });
     }
     let newUser: IUser | null = null;
     try {
@@ -87,7 +87,7 @@ export const addStudent = async (req: Request, res: Response): Promise<any> => {
         const existingUser = await User.findOne({ username });
 
         if (existingUser || existingStudent) {
-            return res.status(400).send({ ...response, error: 'Email already exists for another user.' });
+            return res.status(400).json({ ...response, error: 'Email already exists for another user.' });
         }
         const hashedPassword = await bcrypt.hash(password, 10);
         const userId = uuidv4();
@@ -98,26 +98,28 @@ export const addStudent = async (req: Request, res: Response): Promise<any> => {
         await newStudent.save();
 
         const { userId: studentUserId, ...studentDetails } = newStudent.toObject();
-        res.send({ ...response, success: true, message: 'Student added successfully.', data: { student: studentDetails } });
+        res.status(201).json({ ...response, success: true, message: 'Student added successfully.', data: { student: studentDetails } });
     } catch (error) {
         if (newUser) {
             await newUser.deleteOne();
         }
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
 export const deleteStudent = async (req: Request, res: Response): Promise<any> => {
     const { username } = req.body;
-    
+    if(!username) {
+        return res.status(400).json({ ...response, error: 'Username required' });
+    }
     try {
         const student = await Student.findOne({ _id: username });
         if (!student) {
-            return res.status(404).send({ ...response, error: 'Student not found.' });
+            return res.status(404).json({ ...response, error: 'Student not found.' });
         }
         const user = await User.findOne({ _id: student.userId });
         if (!user) {
-            return res.status(404).send({ ...response, error: 'User not found.' });
+            return res.status(404).json({ ...response, error: 'User not found.' });
         }
         // Remove: (in student collection)
         // feedback: [{ type: Schema.Types.ObjectId, ref: 'Feedback' }],
@@ -133,16 +135,16 @@ export const deleteStudent = async (req: Request, res: Response): Promise<any> =
 
         await User.findByIdAndDelete(user._id);
         await Student.findByIdAndDelete(student._id);
-        res.send({ ...response, success: true, message: 'Student deleted successfully.', student });
+        res.status(200).json({ ...response, success: true, message: 'Student deleted successfully.', data: student });
     } catch (error) {
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
 export const addAdmin = async (req: Request, res: Response): Promise<any> => {
     const { email, password, name } = req.body;
     if (!email || !password || !name) {
-        return res.status(400).send({ ...response, error: 'Missing required fields.' });
+        return res.status(400).json({ ...response, error: 'Missing required fields.' });
     }
     let newUser: IUser | null = null;
     try {
@@ -150,7 +152,7 @@ export const addAdmin = async (req: Request, res: Response): Promise<any> => {
         const existingUser = await User.findOne({ username: email });
 
         if (existingUser || existingAdmin) {
-            return res.status(400).send({ ...response, error: 'Email already exists for another user.' });
+            return res.status(400).json({ ...response, error: 'Email already exists for another user.' });
         }
         const userId = uuidv4();
         const hashedPassword = await bcrypt.hash(password, 10);
@@ -161,35 +163,37 @@ export const addAdmin = async (req: Request, res: Response): Promise<any> => {
         await newAdmin.save();
 
         const { userId: adminUserId, ...adminDetails } = newAdmin.toObject();
-        res.send({ ...response, success: true, message: 'Admin added successfully.', data: { admin: adminDetails } });
+        res.status(201).json({ ...response, success: true, message: 'Admin added successfully.', data: { admin: adminDetails } });
     } catch (error) {
         if (newUser) {
             await newUser.deleteOne();
         }
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
 export const deleteAdmin = async (req: Request, res: Response): Promise<any> => {
     const { email } = req.body;
-    
+    if(!email) {
+        return res.status(400).json({ ...response, error: 'Email required' });
+    }
     try {
         const admin = await Admin.findOne({ email });
         if (!admin) {
-            return res.status(404).send({ ...response, error: 'Admin not found.' });
+            return res.status(404).json({ ...response, error: 'Admin not found.' });
         }
         const user = await User.findOne({ _id: admin.userId });
         if (!user) {
-            return res.status(404).send({ ...response, error: 'User not found.' });
+            return res.status(404).json({ ...response, error: 'User not found.' });
         }
         if(user.username === 'root') {
-            return res.status(403).send({ ...response, error: 'Cannot delete root admin.' });
+            return res.status(403).json({ ...response, error: 'Cannot delete root admin.' });
         }
         await User.findByIdAndDelete(user._id);
         await Admin.findByIdAndDelete(admin._id);
-        res.status(200).send({ ...response, success: true, message: 'Admin deleted successfully.', admin });
+        res.status(200).json({ ...response, success: true, message: 'Admin deleted successfully.', data: admin });
     } catch (error) {
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
@@ -197,29 +201,31 @@ export const deleteAdmin = async (req: Request, res: Response): Promise<any> => 
 export const addCompany = async (req: Request, res: Response): Promise<any> => {
     const { name, tags } = req.body;
     if (!name) {
-        return res.status(400).send({ ...response, error: 'Missing required fields.' });
+        return res.status(400).json({ ...response, error: 'Missing required fields.' });
     }
     try {
         const existingCompany = await Company.findOne({ name });
         if (existingCompany) {
-            return res.status(400).send({ ...response, error: 'Company already exists.' });
+            return res.status(400).json({ ...response, error: 'Company already exists.' });
         }
         const newCompany = new Company({ name, tags: tags || [] });
         await newCompany.save();
-        res.send({ ...response, success: true, message: 'Company added successfully.', data: { company: newCompany } });
+        res.status(201).json({ ...response, success: true, message: 'Company added successfully.', data: { company: newCompany } });
     } catch (error) {
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
 
 // review 
 export const deleteCompany = async (req: Request, res: Response): Promise<any> => {
     const { name } = req.body;
-
+    if(!name) {
+        return res.status(400).json({ ...response, error: 'Company name required' });
+    }
     try {
         const company = await Company.findOne({ name });
         if (!company) {
-            return res.status(404).send({ ...response, error: 'Company not found.' });
+            return res.status(404).json({ ...response, error: 'Company not found.' });
         }
 
         // Remove HRs associated with the company
@@ -233,9 +239,8 @@ export const deleteCompany = async (req: Request, res: Response): Promise<any> =
 
         // Remove the company
         await Company.findByIdAndDelete(company._id);
-
-        res.send({ ...response, success: true, message: 'Company deleted successfully.', company });
+        res.status(200).json({ ...response, success: true, message: 'Company deleted successfully.', data: company });
     } catch (error) {
-        res.status(500).send({ ...response, error });
+        res.status(500).json({ ...response, error });
     }
 }
